Move warnings to top-level
