+++ l2m2/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "0.0.41"
+__version__ = "0.0.41.next"
